+++ routes/club/club.js
@@ -8,12 +8,6 @@ router.use((req, res, next) => {
   res.locals.followerCount = 0;
   res.locals.followingCount = 0;
   res.locals.followerIdList = [];
-  // res.locals.user = req.user;
-  // res.locals.followerCount = req.user ? req.user.Followers.length : 0;
-  // res.locals.followingCount = req.user ? req.user.Followings.length : 0;
-  // res.locals.followerIdList = req.user
-  // ? req.user.Followings.map((f) => f.id)
-  // : [];
   next();
 });
 
@@ -52,8 +46,8 @@ router.use((req, res, next) => {
 
 router.get("/", async (req, res, next) => {
   try {
-    const clubs = await Club.findAll({ 
-      include: { 
+    const clubs = await Club.findAll({
+      include: {
         model: User,
         attribute: ['id', 'nick'],
       },
