+++ views/board/edit-community.html
@@ -44,7 +44,6 @@
         </form>
       </body>
 </section>
-
 {% endblock %}
 
 {% block script %}
