+++ views/mypage/mypage.html
@@ -13,8 +13,6 @@
     <div class="count follower-count">{{followerCount}}</div>
   </div>
 
-  <input id="my-id" type="hidden" value="{{user.id}}" />
-
   <div class="followings half">
     <h2>팔로잉 목록</h2>
     {% if user.Followings %} {% for following in user.Followings %}
